requirements.txt
aiohttp==3.8.1
sqlalchemy==1.4.41
pyyaml==6.0
asyncpg==0.26.0
aiohttp_session==2.11.0
alembic==1.8.1
cryptography==38.0.1
pycparser==2.21
aiohttp_session==2.11.0
cryptography==38.0.1
cffi==1.15.1
pycparser==2.21
aiohttp-apispec==2.2.3

